tirithel/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tirithel"
version = "0.1.0"
description = "AI Support UI Learning Tool - learns software navigation from support sessions"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    "anthropic>=0.39.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.2.0",
    "opencv-python-headless>=4.9.0",
    "numpy>=1.26.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["tirithel*"]
